+++ R/elnet.R
@@ -18,7 +18,7 @@ elnet=function(x,is.sparse,ix,jx,y,weights,offset,type.gaussian=c("covariance","
     y=y-offset
   }
 ### compute the null deviance
-  ybar=weighted.mean(y,weights)
+  ybar=if(intr)weighted.mean(y,weights)else 0
   nulldev=sum(weights* (y-ybar)^2)
 if(nulldev==0)stop("y is constant; gaussian glmnet fails at standardization step")