+++ R/glmnet.R
@@ -8,8 +8,12 @@
 #'
 #' The sequence of models implied by \code{lambda} is fit by coordinate
 #' descent. For \code{family="gaussian"} this is the lasso sequence if
-#' \code{alpha=1}, else it is the elasticnet sequence.  For the other families,
-#' this is a lasso or elasticnet regularization path for fitting the
+#' \code{alpha=1}, else it is the elasticnet sequence.
+#'
+#' From version 4.0 onwards, glmnet supports both the original built-in families,
+#' as well as \emph{any} family object as used by `stats:glm()`.
+#' The built in families are specifed via a character string. For all families,
+#' the object produced is a lasso or elasticnet regularization path for fitting the
 #' generalized linear regression paths, by maximizing the appropriate penalized
 #' log-likelihood (partial likelihood for the "cox" model). Sometimes the
 #' sequence is truncated before \code{nlambda} values of \code{lambda} have
@@ -29,8 +33,8 @@
 #' before computing its lambda sequence (and then unstandardizes the resulting
 #' coefficients); if you wish to reproduce/compare results with other software,
 #' best to supply a standardized y. The coefficients for any predictor
-#' variables with zero variance are set to zero for all values of lambda.  The
-#' latest two features in glmnet are the \code{family="mgaussian"} family and
+#' variables with zero variance are set to zero for all values of lambda.
+#' Two useful additional families  are the \code{family="mgaussian"} family and
 #' the \code{type.multinomial="grouped"} option for multinomial fitting. The
 #' former allows a multi-response gaussian model to be fit, using a "group
 #' -lasso" penalty on the coefficients for each variable. Tying the responses
@@ -44,7 +48,15 @@
 #' with censored times, we assume the censored times occurred just
 #' \emph{before} the death times in computing the Breslow approximation; if
 #' users prefer the usual convention of \emph{after}, they can add a small
-#' number to all censoring times to achieve this effect.  If \code{relax=TRUE}
+#' number to all censoring times to achieve this effect.
+#'
+#' Version 4.0 and later allows for the family argument to be a S3 class `"family"` object
+#' (a list of functions and expressions).
+#' This opens the door to a wide variety of additional models. For example
+#' `family=binomial(link=cloglog)` or `family=negative.binomial(theta=1.5)` (from the MASS library).
+#' Note that the code runs faster for the built-in families.
+#'
+#' If \code{relax=TRUE}
 #' a duplicate sequence of models is produced, where each active set in the
 #' elastic-net path is refit without regularization. The result of this is a
 #' matching \code{"glmnet"} object which is stored on the original object in a
@@ -72,7 +84,8 @@
 #' right censored. The function \code{Surv()} in package \pkg{survival}
 #' produces such a matrix. For \code{family="mgaussian"}, \code{y} is a matrix
 #' of quantitative responses.
-#' @param family Response type (see above)
+#' @param family Response type (see above). Either a character string representing
+#' one of the built-in families, or else a `glm()` family object.
 #' @param weights observation weights. Can be total counts if responses are
 #' proportion matrices. Default is 1 for each observation
 #' @param offset A vector of length \code{nobs} that is included in the linear
@@ -212,8 +225,9 @@
 #' Statistics Technical Report}\cr \url{https://arxiv.org/abs/1707.08692}\cr
 #' Hastie, T., Tibshirani, Robert, Tibshirani, Ryan (2019) \emph{Extended
 #' Comparisons of Best Subset Selection, Forward Stepwise Selection, and the
-#' Lasso}\cr \emph{Glmnet Vignette}
-#' \url{https://web.stanford.edu/~hastie/glmnet/glmnet_alpha.html}
+#' Lasso}\cr
+#' \emph{Glmnet webpage with four vignettes}
+#' \url{https://glmnet.stanford.edu}
 #' @keywords models regression
 #' @examples
 #'
@@ -234,8 +248,9 @@
 #' plot(fit1m, type.coef = "2norm")
 #'
 #' # binomial
-#' g2 = sample(1:2, 100, replace = TRUE)
+#' g2 = sample(c(0,1), 100, replace = TRUE)
 #' fit2 = glmnet(x, g2, family = "binomial")
+#' fit2n = glmnet(x, g2, family = binomial(link=cloglog))
 #' fit2r = glmnet(x,g2, family = "binomial", relax=TRUE)
 #' fit2rp = glmnet(x,g2, family = "binomial", relax=TRUE, path=TRUE)
 #'
@@ -292,8 +307,30 @@
 #' system.time(fit2n <- glmnet(x, y))
 #'
 #' @export glmnet
-glmnet=function(x,y,family=c("gaussian","binomial","poisson","multinomial","cox","mgaussian"),weights,offset=NULL,alpha=1.0,nlambda=100,lambda.min.ratio=ifelse(nobs<nvars,1e-2,1e-4),lambda=NULL,standardize=TRUE,intercept=TRUE,thresh=1e-7,dfmax=nvars+1,pmax=min(dfmax*2+20,nvars),exclude,penalty.factor=rep(1,nvars),lower.limits=-Inf,upper.limits=Inf,maxit=100000,type.gaussian=ifelse(nvars<500,"covariance","naive"),type.logistic=c("Newton","modified.Newton"),standardize.response=FALSE,type.multinomial=c("ungrouped","grouped"),relax=FALSE,trace.it=0,...){
+glmnet=function(x,y,family=c("gaussian","binomial","poisson","multinomial","cox","mgaussian"),weights=NULL,offset=NULL,alpha=1.0,nlambda=100,lambda.min.ratio=ifelse(nobs<nvars,1e-2,1e-4),lambda=NULL,standardize=TRUE,intercept=TRUE,thresh=1e-7,dfmax=nvars+1,pmax=min(dfmax*2+20,nvars),exclude=NULL,penalty.factor=rep(1,nvars),lower.limits=-Inf,upper.limits=Inf,maxit=100000,type.gaussian=ifelse(nvars<500,"covariance","naive"),type.logistic=c("Newton","modified.Newton"),standardize.response=FALSE,type.multinomial=c("ungrouped","grouped"),relax=FALSE,trace.it=0,...){
 
+    this.call=match.call()
+### Need to do this first so defaults in call can be satisfied
+    np=dim(x)
+    ##check dims
+    if(is.null(np)|(np[2]<=1))stop("x should be a matrix with 2 or more columns")
+    nobs=as.integer(np[1])
+    nvars=as.integer(np[2])
+
+### See whether its a call to glmnet or to glmnet.path, based on family arg
+    if(!is.character(family)){
+        ## new.call=this.call
+        ## new.call[[1]]=as.name("glmnet.path")
+        ## fit=eval(new.call, parent.frame())
+        if(missing(thresh))thresh=1e-10
+        fit=glmnet.path(x,y,weights,lambda,nlambda,lambda.min.ratio,alpha,offset,family,
+                        standardize,intercept,thresh=thresh,maxit,penalty.factor,exclude,lower.limits,
+                        upper.limits,trace.it=trace.it)
+        fit$call=this.call
+    }
+    else
+    {
+### Must have been a call to old glmnet
 ### Prepare all the generic arguments, then hand off to family functions
   family=match.arg(family)
   if(alpha>1){
@@ -305,17 +342,10 @@ glmnet=function(x,y,family=c("gaussian","binomial","poisson","multinomial","cox"
     alpha=0
   }
   alpha=as.double(alpha)
-
-  this.call=match.call()
   nlam=as.integer(nlambda)
   y=drop(y) # we dont like matrix responses unless we need them
-  np=dim(x)
-   ###check dims
-  if(is.null(np)|(np[2]<=1))stop("x should be a matrix with 2 or more columns")
- nobs=as.integer(np[1])
-  if(missing(weights))weights=rep(1,nobs)
+  if(is.null(weights))weights=rep(1,nobs)
   else if(length(weights)!=nobs)stop(paste("number of elements in weights (",length(weights),") not equal to the number of rows of x (",nobs,")",sep=""))
-  nvars=as.integer(np[2])
   dimy=dim(y)
   nrowy=ifelse(is.null(dimy),length(y),dimy[1])
     if(nrowy!=nobs)stop(paste("number of observations in y (",nrowy,") not equal to the number of rows of x (",nobs,")",sep=""))
@@ -323,7 +353,7 @@ glmnet=function(x,y,family=c("gaussian","binomial","poisson","multinomial","cox"
   if(is.null(vnames))vnames=paste("V",seq(nvars),sep="")
   ne=as.integer(dfmax)
     nx=as.integer(pmax)
-    if(missing(exclude))exclude=integer(0)
+    if(is.null(exclude))exclude=integer(0)
     if(any(penalty.factor==Inf)){
         exclude=c(exclude,seq(nvars)[penalty.factor==Inf])
         exclude=sort(unique(exclude))
@@ -426,11 +456,11 @@ glmnet=function(x,y,family=c("gaussian","binomial","poisson","multinomial","cox"
 fit$call=this.call
   fit$nobs=nobs
     class(fit)=c(class(fit),"glmnet")
+    }
     if(relax)
         relax.glmnet(fit, x=x,y=y,weights=weights,offset=offset,
-                     lower.limits=lower.limits,upper.limits=upper.limits,
+                     lower.limits=lower.limits,upper.limits=upper.limits,penalty.factor=penalty.factor,
                      check.args=FALSE,...)
     else
         fit
-
 }