+++ R/glmnet.measures.R
@@ -13,13 +13,13 @@
 #' @keywords models
 #' @export glmnet.measures
 glmnet.measures=function(family=c("all","gaussian", "binomial", "poisson",
-                                  "multinomial", "cox", "mgaussian")){
+                                  "multinomial", "cox", "mgaussian","GLM")){
     family=match.arg(family)
     type.measures = c("mse", "deviance", "class", "auc", "mae",
                       "C")
     family.ch = list( gaussian = c(1, 5), binomial = c(2,
                                                           3, 4, 1, 5), poisson = c(2, 1, 5), cox = c(2, 6),
-                     multinomial = c(2, 3, 1, 5), mgaussian = c(1, 5))
+                     multinomial = c(2, 3, 1, 5), mgaussian = c(1, 5),GLM=c(2,1,5))
     if(family=="all")lapply(family.ch,function(x)type.measures[x])
     else         type.measures[family.ch[[family]]]